Phase 6.1: extract NullRawJSON into shared server/sqlhelpers
Two contexts (detection, response) shipped private NullRawJSON
implementations. Consolidate into a single canonical home so the
behavior + tests don't drift across copies.

- server/sqlhelpers/: new package, owns NullRawJSON (Scan/Value
  + Marshal/Unmarshal) plus PBT round-trip tests.
- detection/api/types.go: NullRawJSON becomes a type alias to
  sqlhelpers.NullRawJSON; public surface (and rules.api re-export)
  stays byte-identical.
- response/internal/mysql/store.go: drops the private nullRawJSON
  type; commandRow.Result switches to sqlhelpers.NullRawJSON.
  api.Command.Result wire shape (json.RawMessage at the api edge)
  stays unchanged via toAPI's existing conversion.
- detection/api/api_test.go: drop NullRawJSON tests (now in
  sqlhelpers/); JSONStringSlice + Event + validation-error tests
  remain.
